관리자 인증 검증 순환 참조 제거
