+++ src/main/java/com/yd/vibecode/global/security/SecurityConfig.java
@@ -11,7 +11,6 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
@@ -19,7 +18,7 @@ import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.CorsUtils;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import com.yd.vibecode.domain.auth.domain.service.AdminService;
+import com.yd.vibecode.domain.auth.domain.service.AdminAccessAuthValidator;
 import com.yd.vibecode.domain.auth.domain.service.RefreshTokenService;
 import com.yd.vibecode.domain.auth.domain.service.TokenWhitelistService;
 import com.yd.vibecode.global.config.properties.CorsProperties;
@@ -40,7 +39,7 @@ public class SecurityConfig {
 	private final ExcludeAuthPathProperties excludeAuthPathProperties;
 	private final RefreshTokenService refreshTokenService;
 	private final TokenWhitelistService tokenWhitelistService;
-	private final AdminService adminService;
+	private final AdminAccessAuthValidator adminAccessAuthValidator;
 	private final CorsProperties corsProperties;
 
 	@Bean
@@ -124,13 +123,8 @@ public class SecurityConfig {
                 excludeAuthPathProperties,
                 refreshTokenService,
                 tokenWhitelistService,
-                adminService
+                adminAccessAuthValidator
         );
-    }
-    
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
     }
 
     private void writeUnauthorizedResponse(HttpServletResponse response) throws IOException {
