+++ src/test/java/com/yd/vibecode/domain/auth/domain/service/AdminServiceTest.java
@@ -20,8 +20,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.test.util.ReflectionTestUtils;
-
 @ExtendWith(MockitoExtension.class)
 class AdminServiceTest {
 
@@ -81,63 +79,6 @@ class AdminServiceTest {
 
         assertThat(admin.getLastLoginAt()).isNotNull();
         verify(adminRepository, never()).save(any(Admin.class));
-    }
-
-    @Test
-    @DisplayName("삭제된 관리자는 access token 인증에서 차단된다")
-    void validateActiveForAuthentication_deletedAdmin_throws() {
-        Admin admin = Admin.builder()
-                .adminNumber("admin123")
-                .email("admin@example.com")
-                .passwordHash("encoded")
-                .isActive(false)
-                .build();
-        ReflectionTestUtils.setField(admin, "id", 1L);
-        admin.delete();
-
-        given(adminRepository.findById(1L)).willReturn(Optional.of(admin));
-
-        RestApiException exception = assertThrows(
-                RestApiException.class,
-                () -> adminService.validateActiveForAuthentication(1L));
-
-        assertThat(exception.getErrorCode()).isEqualTo(AuthErrorStatus.ADMIN_ACCOUNT_INACTIVE.getCode());
-    }
-
-    @Test
-    @DisplayName("비활성 관리자는 access token 인증에서 차단된다")
-    void validateActiveForAuthentication_inactiveAdmin_throws() {
-        Admin admin = Admin.builder()
-                .adminNumber("admin123")
-                .email("admin@example.com")
-                .passwordHash("encoded")
-                .isActive(false)
-                .build();
-        ReflectionTestUtils.setField(admin, "id", 2L);
-
-        given(adminRepository.findById(2L)).willReturn(Optional.of(admin));
-
-        RestApiException exception = assertThrows(
-                RestApiException.class,
-                () -> adminService.validateActiveForAuthentication(2L));
-
-        assertThat(exception.getErrorCode()).isEqualTo(AuthErrorStatus.ADMIN_ACCOUNT_INACTIVE.getCode());
-    }
-
-    @Test
-    @DisplayName("활성 관리자는 access token 인증 검증을 통과한다")
-    void validateActiveForAuthentication_activeAdmin_passes() {
-        Admin admin = Admin.builder()
-                .adminNumber("admin123")
-                .email("admin@example.com")
-                .passwordHash("encoded")
-                .isActive(true)
-                .build();
-        ReflectionTestUtils.setField(admin, "id", 3L);
-
-        given(adminRepository.findById(3L)).willReturn(Optional.of(admin));
-
-        assertDoesNotThrow(() -> adminService.validateActiveForAuthentication(3L));
     }
 
     @Test
